\name{DurocMk}
\alias{DurocMk}
\docType{data}
\title{
 SNPs information in Duroc from Porcine SNP60 Beadchips 
}
\description{
 This data set gives the marker information from Porcine SNP60 Beadchips
}
\usage{DurocMk}
\format{
  A data frame with 62551 observations on the following 4 variables.
  \describe{
    \item{\code{MkName}}{a character vector}
    \item{\code{Chromosome}}{a character vector}
    \item{\code{Position}}{a numeric vector}
    \item{\code{Anno}}{a character vector}
  }
}

\source{
\url{http://asia.ensembl.org/index.html}
}

\examples{
data(DurocMk)
}
\keyword{datasets}
