\name{F2Recomb}
\alias{F2Recomb}
\docType{data}
\title{
SNPs recombination information from F2 breed 
}
\description{
60K pig SNPs recombination information from a white Duroc x Erhualian F2 resource 
}
\usage{F2Recomb}
\format{
  A data frame with 35063 observations on the following 2 variables.
  \describe{
    \item{\code{recomb}}{a numeric vector}
    \item{\code{MkName}}{a character vector}
  }
}

\references{
Ren J. et al. (2006) WCGALP 8th, 13-1.
}

\examples{
data(F2Recomb)
}
\keyword{datasets}
