\name{MCV_pval}
\alias{MCV_pval}
\docType{data}
\title{
 P-value information of mean corpuscular volume
}
\description{
This data set gives the P-value information of mean corpuscular volume using in example
}
\usage{MCV_pval}
\format{
  A data frame with 39787 observations on the following 2 variables.
  \describe{
    \item{\code{MkName}}{a character vector}
    \item{\code{Pval}}{a numeric vector}
  }
}
\references{
Zhang F. et al. (2014) BMC Genet. 15, 41}
\examples{
data(MCV_pval)
}
\keyword{datasets}
