\name{MeanLD}
\alias{MeanLD}
\docType{data}
\title{
Pair-wise LD information 
}
\description{
Pair-wise LD information from a F2 resource pedigree for 1020 individuals and Sutai breed including 436 individuals
}
\usage{MeanLD}
\format{
  A data frame with 1980089 observations on the following 4 variables.
  \describe{
    \item{\code{CHR_A}}{a character vector}
    \item{\code{SNP_A}}{a character vector}
    \item{\code{SNP_B}}{a character vector}
    \item{\code{R2}}{a numeric vector}
  }
}

\references{
Ren J. et al. (2006) WCGALP 8th, 13-1.
Ma J.W. et al. (2013) PloS one 8, e64047.
}
\examples{
data(MeanLD)
}
\keyword{datasets}
