\name{SutaiRecomb}
\alias{SutaiRecomb}
\docType{data}
\title{
SNPs recombination information from Sutai breed 
}
\description{
60K pig SNPs recombination information from Sutai pig 
}
\usage{SutaiRecomb}
\format{
  A data frame with 39786 observations on the following 2 variables.
  \describe{
    \item{\code{recomb}}{a numeric vector}
    \item{\code{MkName}}{a character vector}
  }
}
\references{
Ma J.W. et al. (2013) PloS one 8, e64047
}
\examples{
data(SutaiRecomb)
}
\keyword{datasets}
