\name{WzsMk}
\alias{WzsMk}
\docType{data}
\title{
SNPs information in WuZhiShan from Porcine SNP60 Beadchips 
}
\description{
 This data set gives the marker information from Porcine SNP60 Beadchips
}
\usage{WzsMk}
\format{
  A data frame with 62551 observations on the following 4 variables.
  \describe{
   \item{\code{MkName}}{a character vector}
    \item{\code{Chromosome}}{a character vector}
    \item{\code{Position}}{a numeric vector}
    \item{\code{Anno}}{a character vector}
  }
}


\references{
Fang X.D. et al. (2012) GigaScience 1, 16.
}
\examples{
data(WzsMk)
}
\keyword{datasets}
