% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotRegion}
\alias{plotRegion}
\title{Regional Plot Association Results of Pig}
\usage{
plotRegion(pval, plotChr, plotRegion, geneRegion, isLD = TRUE, LDresult, geno,
  isRecomb = TRUE, Recomb, isAnno = TRUE, main = "", ref = "Wzs", Map,
  Gff, xlab)
}
\arguments{
\item{pval}{A R object.the input data of GWAS result constructed by at least 2 columns,
P-value and SNP identity in Illumina system. The two colnames of pval must be
"MkName" and "Pval".}

\item{plotChr}{a character:the chromosome to plot. such as "1", "2", "X" and "Y".}

\item{plotRegion}{a numeric vector containing the start position and the end position of plot.
The unit is "Mb".}

\item{geneRegion}{a numeric vector containing the start position and the end position of gene structure in bottom panel.
The unit is "Mb".}

\item{isLD}{a logical value indicating whether plots the LD information}

\item{LDresult}{A character value or a dataframe. As a dataframe, the data of linkage disequilibrium (LD) consisting of the maker name of other SNPs in plotting Region and
the R2 associated with the top SNP. The two colnames are "SNP_B" and "R2", respectively.
As a character value, "Other" for other species.}

\item{geno}{A dataframe for genotype data. If we want to use own LD pattern, the package will automatically call "genetics" package to calculated LD.}

\item{isRecomb}{a logical value indicating whether plots the recombination information}

\item{Recomb}{A dataframe contains the SNPs name and the recombination information, of which colnames are "MkName" and "recomb", respectively.
SutaiRecomb and F2Recomb are two built-in databases representing recombination of Sutai breed and F2 breed.}

\item{isAnno}{a logical value indicating whether plots every SNP annotaion information}

\item{main}{The main title (on top) using font and size (character expansion) par("font.main") and color par("col.main").}

\item{ref}{the name of the reference genome. There are three different genome:
the widespread commercial breed: "Duroc" (Sscrofa10.2) or Chinese indigenous breed: Wzs (Wuzhishan) or other species: "Other".}

\item{Map}{The Map information for other species (used only if ref is specified for "Other")}

\item{xlab}{a title for the x axis.}
}
\value{
a figure
}
\description{
This is a plotting tool to simultaneously display the association p-value,
             the LD, recombination rate and annotated gene in region of interesting in pig.
             The interpretation and visualization tool of association results will
             contribute to further identification of causal mutations.
}
\details{
nothing
}
\examples{
#plotting for pig
png("MCV.png", width=3200, height=2200,res=300) # this is the most appropriate size and resolution
plotRegion(MCV_pval)
dev.off()
\dontrun{
#plotting for other species
plotRegion(MCV_pval,ref="Other",LDresult=LDresult,Recomb=Recomb,Map=Map,Gff=Gff)
}
}
\author{
Feng Zhang and Zhiyan Zhang
}

